Splits Spinner, Router to own components
