+++ app/components/Spinner.tsx
@@ -1,0 +1,11 @@
+import '../styles/spinner.css'
+
+export const Spinner = () => (
+  <div className="spinner">
+    <div className="rect1"></div>
+    <div className="rect2"></div>
+    <div className="rect3"></div>
+    <div className="rect4"></div>
+    <div className="rect5"></div>
+  </div>
+)
