+++ app/styles/spinner.css
@@ -1,33 +1,41 @@
 .spinner {
-    margin: 2px auto;
+  margin: 2px auto;
+  height: 1rem;
+  display: inline-block;
+}
+
+.spinner > div {
+  background-color: #555;
+  height: 80%;
+  width: 3px;
+  display: inline-block;
+
+  animation: sk-stretchdelay 1.2s infinite ease-in-out;
+}
+
+.spinner .rect2 {
+  animation-delay: -1.1s;
+}
+
+.spinner .rect3 {
+  animation-delay: -1s;
+}
+
+.spinner .rect4 {
+  animation-delay: -0.9s;
+}
+
+.spinner .rect5 {
+  animation-delay: -0.8s;
+}
+
+@keyframes sk-stretchdelay {
+  0%,
+  40%,
+  100% {
+    transform: scaleY(0.4);
   }
-  
-  .spinner > div {
-    background-color: #333;
-    height: 80%;
-    width: 3px;
-    display: inline-block;
-    
-    animation: sk-stretchdelay 1.2s infinite ease-in-out;
+  20% {
+    transform: scaleY(1);
   }
-  
-  .spinner .rect2 {
-    animation-delay: -1.1s;
-  }
-  
-  .spinner .rect3 {
-    animation-delay: -1.0s;
-  }
-  
-  .spinner .rect4 {
-    animation-delay: -0.9s;
-  }
-  
-  .spinner .rect5 {
-    animation-delay: -0.8s;
-  }
-  
-  @keyframes sk-stretchdelay {
-    0%, 40%, 100% { transform: scaleY(0.4) }  
-    20% { transform: scaleY(1.0) }
-  }
+}
